Changed count to long so it is not necessary to cast
Stream::count returns a long

+++ src/main/java/com/example/application/views/chat/ChatView.java
@@ -41,7 +41,7 @@ public class ChatView extends VerticalLayout {
             this.channel = channel;
         }
 
-        void setCount(int count) {
+        void setCount(long count) {
             this.count.setText("" + count);
         }
